Add Next.js configuration and refactor Birthday component
- Introduced a new Next.js configuration file to enable strict mode.
- Refactored the Birthday component into a separate file for better organization and dynamic import.
- Updated the main page to utilize the new Birthday component with client-side rendering.
- Removed unused state and effects from the Birthday component for cleaner code.
