+++ frontend/next.config.mjs
@@ -1,0 +1,7 @@
+/** @type {import('next').NextConfig} */
+const nextConfig = {
+  // Configure Next.js to run in a more compatible mode
+  reactStrictMode: true,
+};
+
+export default nextConfig;
